Add files via upload
Here is another Python program that provides a simple task management function. The user can add, view and delete tasks. The program saves the tasks in a text file.
This program allows the user to add, view and delete tasks. The tasks are saved in a text file tasks.txt. The program provides the following functions:

"Add a task": the user can enter a new task to add it to the list.
"List tasks": The user gets a numbered list of all tasks.
"Delete a task": the user can enter the task number to delete it from the list.
"Quit": exits the program.
As before, you can download this program, customize it and upload it to GitHub to share it with others. Make sure you create the tasks.txt file in the same directory as the program so that the tasks can be saved correctly.

You can continue to develop the program and add additional features to make it more comprehensive. Have fun programming!